+++ src/lib/services/blockService.ts
@@ -2,7 +2,7 @@ import type { UseQueryResult } from "@tanstack/react-query";
 import { useQuery } from "@tanstack/react-query";
 import { useCallback } from "react";
 
-import { useCelatoneApp } from "lib/app-provider";
+import { CELATONE_QUERY_KEYS, useCelatoneApp } from "lib/app-provider";
 import {
   getLatestBlockInfoQueryDocument,
   getBlockDetailsByHeightQueryDocument,
@@ -53,7 +53,10 @@ export const useBlocklistQuery = (
     [indexerGraphClient, currentChainId, limit, offset]
   );
 
-  return useQuery(["blocks", indexerGraphClient, limit, offset], queryFn);
+  return useQuery(
+    [CELATONE_QUERY_KEYS.BLOCKS, indexerGraphClient, limit, offset],
+    queryFn
+  );
 };
 
 export const useBlockCountQuery = (): UseQueryResult<number> => {
@@ -66,10 +69,13 @@ export const useBlockCountQuery = (): UseQueryResult<number> => {
     [indexerGraphClient]
   );
 
-  return useQuery(["block_count", indexerGraphClient], queryFn);
+  return useQuery(
+    [CELATONE_QUERY_KEYS.BLOCK_COUNT, indexerGraphClient],
+    queryFn
+  );
 };
 
-export const useBlockDetailsQuery = (
+export const useBlockInfoQuery = (
   height: string
 ): UseQueryResult<BlockDetails | null> => {
   const { currentChainId } = useCelatoneApp();
@@ -105,11 +111,15 @@ export const useBlockDetailsQuery = (
     [indexerGraphClient, currentChainId, height]
   );
 
-  return useQuery(["block_details", indexerGraphClient, height], queryFn, {
-    enabled: isBlock(height),
-    retry: false,
-    refetchOnWindowFocus: false,
-  });
+  return useQuery(
+    [CELATONE_QUERY_KEYS.BLOCK_INFO, indexerGraphClient, height],
+    queryFn,
+    {
+      enabled: isBlock(height),
+      retry: false,
+      refetchOnWindowFocus: false,
+    }
+  );
 };
 
 export const useLatestBlockInfo = (): UseQueryResult<LatestBlock> => {
@@ -125,5 +135,8 @@ export const useLatestBlockInfo = (): UseQueryResult<LatestBlock> => {
     [indexerGraphClient]
   );
 
-  return useQuery(["latest_block_info", indexerGraphClient], queryFn);
+  return useQuery(
+    [CELATONE_QUERY_KEYS.LATEST_BLOCK_INFO, indexerGraphClient],
+    queryFn
+  );
 };