+++ src/lib/services/codeService.ts
@@ -3,7 +3,11 @@ import type { UseQueryResult } from "@tanstack/react-query";
 import { useQuery } from "@tanstack/react-query";
 import { useCallback } from "react";
 
-import { useCelatoneApp, useWasmConfig } from "lib/app-provider";
+import {
+  CELATONE_QUERY_KEYS,
+  useCelatoneApp,
+  useWasmConfig,
+} from "lib/app-provider";
 import {
   getCodeDataByCodeId,
   getCodeListByIDsQueryDocument,
@@ -45,7 +49,7 @@ export const useCodeListQuery = (): UseQueryResult<CodeInfo[]> => {
       );
   }, [indexerGraphClient]);
 
-  return useQuery(["recents_codes", indexerGraphClient], queryFn);
+  return useQuery([CELATONE_QUERY_KEYS.CODES, indexerGraphClient], queryFn);
 };
 
 export const useCodeListByWalletAddress = (
@@ -76,7 +80,11 @@ export const useCodeListByWalletAddress = (
   }, [walletAddr, indexerGraphClient]);
 
   return useQuery(
-    ["code_list_by_user", walletAddr, indexerGraphClient],
+    [
+      CELATONE_QUERY_KEYS.CODES_BY_WALLET_ADDRESS,
+      walletAddr,
+      indexerGraphClient,
+    ],
     queryFn,
     {
       keepPreviousData: true,
@@ -111,10 +119,14 @@ export const useCodeListByCodeIds = (
       );
   }, [ids, indexerGraphClient]);
 
-  return useQuery(["code_list_by_ids", ids, indexerGraphClient], queryFn, {
-    keepPreviousData: true,
-    enabled: !!ids,
-  });
+  return useQuery(
+    [CELATONE_QUERY_KEYS.CODES_BY_IDS, ids, indexerGraphClient],
+    queryFn,
+    {
+      keepPreviousData: true,
+      enabled: !!ids,
+    }
+  );
 };
 
 export const useCodeDataByCodeId = (
@@ -156,9 +168,13 @@ export const useCodeDataByCodeId = (
         };
       });
   }, [codeId, indexerGraphClient]);
-  return useQuery(["code_data_by_id", codeId, indexerGraphClient], queryFn, {
-    enabled: enabled && isCodeId(codeId),
-  });
+  return useQuery(
+    [CELATONE_QUERY_KEYS.CODE_DATA_BY_ID, codeId, indexerGraphClient],
+    queryFn,
+    {
+      enabled: enabled && isCodeId(codeId),
+    }
+  );
 };
 
 export const useCodeListByWalletAddressPagination = (
@@ -197,7 +213,7 @@ export const useCodeListByWalletAddressPagination = (
 
   return useQuery(
     [
-      "code_list_by_wallet_address_pagination",
+      CELATONE_QUERY_KEYS.CODES_BY_WALLET_ADDRESS_PAGINATION,
       indexerGraphClient,
       offset,
       pageSize,
@@ -232,7 +248,11 @@ export const useCodeListCountByWalletAddress = (
   }, [walletAddress, indexerGraphClient]);
 
   return useQuery(
-    ["code_list_count_by_wallet_address", walletAddress, indexerGraphClient],
+    [
+      CELATONE_QUERY_KEYS.CODES_BY_WALLET_ADDRESS_COUNT,
+      walletAddress,
+      indexerGraphClient,
+    ],
     queryFn,
     {
       keepPreviousData: true,